+++ jgclark.DashboardReact/src/react/components/Section.jsx
@@ -11,7 +11,7 @@
 import React, { useState, useEffect } from 'react'
 import type { TSection, TSectionItem } from '../../types.js'
 import useInteractiveProcessing from '../customHooks/useInteractiveProcessing.jsx'
-import { getFeatureFlags } from '../../shared.js'
+import useSectionSortAndFilter from '../customHooks/useSectionSortAndFilter.jsx'
 import CommandButton from './CommandButton.jsx'
 import ItemGrid from './ItemGrid.jsx'
 import { useAppContext } from './AppContext.jsx'
@@ -21,8 +21,8 @@ import { logDebug, logError, JSP } from '@helpers/react/reactDev'
 // Type Definitions
 //--------------------------------------------------------------------------
 type SectionProps = {
-  section: TSection
-}
+  section: TSection,
+};
 
 //--------------------------------------------------------------------------
 // Section Component Definition
@@ -36,8 +36,8 @@ const Section = ({ section }: SectionProps): React$Node => {
   //----------------------------------------------------------------------
   // State
   //----------------------------------------------------------------------
-  const [itemsCopy, setItemsCopy] = useState < Array < TSectionItem >> ([])
-  const [items, setItems] = useState < Array < TSectionItem >> ([])
+  const [items, setItems] = useState<Array<TSectionItem>>([])
+  const [itemsCopy, setItemsCopy] = useState<Array<TSectionItem>>([])
 
   //----------------------------------------------------------------------
   // Effects
@@ -49,14 +49,12 @@ const Section = ({ section }: SectionProps): React$Node => {
     }
 
     if (sharedSettings && section.showSettingName && sharedSettings[section.showSettingName] === false) {
-      logDebug('Section', `Section: ${section.ID} ("${section.name}") is currently filtered out sharedSettings[${section.showSettingName}]=${sharedSettings?.[section.showSettingName]}`)
       return
     }
 
     let sectionItems = section.sectionItems
     if (!sectionItems || sectionItems.length === 0) {
       if (section.ID !== 0) {
-        // logDebug('Section', `Section: ${section.ID} / ${section.name} (${section.sectionCode}) is switched off or has no sectionItems`)
         sectionItems = []
       } else {
         logDebug('Section', `Section 0 doesn't have any sectionItems, so display congrats message`)
@@ -73,7 +71,9 @@ const Section = ({ section }: SectionProps): React$Node => {
   //----------------------------------------------------------------------
   // Hooks
   //----------------------------------------------------------------------
-  useInteractiveProcessing(items, section, itemsCopy, setItemsCopy, reactSettings, setReactSettings, sendActionToPlugin, sharedSettings)
+  const { filteredItems, itemsToShow, filteredOut, limitApplied } = useSectionSortAndFilter(section, items, sharedSettings)
+  
+  useInteractiveProcessing(filteredItems, section, itemsCopy, setItemsCopy, reactSettings, setReactSettings, sendActionToPlugin, sharedSettings)
 
   //----------------------------------------------------------------------
   // Handlers
@@ -85,78 +85,15 @@ const Section = ({ section }: SectionProps): React$Node => {
       ...prevSettings,
       lastChange: `_InteractiveProcessing Click`,
       interactiveProcessing: { sectionName: section.name, currentIPIndex: 0, totalTasks: itemsToShow.length, clickPosition, startingUp: true },
-      dialogData: { isOpen: false, isTask: true, details: {}, clickPosition }
+      dialogData: { isOpen: false, isTask: true, details: {}, clickPosition },
     }))
   }
 
   //----------------------------------------------------------------------
-  // Constants
+  // Render
   //----------------------------------------------------------------------
-  const buttons = section.actionButtons?.map((item, index) => <CommandButton key={index} button={item} />) ?? []
-
-  const { enableInteractiveProcessing } = sharedSettings
-
-  const filterPriorityItems = sharedSettings?.filterPriorityItems ?? false
-  let maxPrioritySeen = 0
-  for (const i of items) {
-    if (i.para?.priority && i.para.priority > maxPrioritySeen) {
-      maxPrioritySeen = i.para.priority
-    }
-  }
-
-  const filteredItems = filterPriorityItems ? items.filter((f) => (f.para?.priority ?? 0) >= maxPrioritySeen) : items.slice()
-  const priorityFilteringHappening = items.length > filteredItems.length
-
-  filteredItems.sort((a, b) => {
-    if (a.para?.startTime && b.para?.startTime) {
-      const startTimeComparison = a.para.startTime.localeCompare(b.para.startTime)
-      if (startTimeComparison !== 0) return startTimeComparison
-    } else if (a.para?.startTime) {
-      return -1
-    } else if (b.para?.startTime) {
-      return 1
-    }
-
-    if (a.para?.endTime && b.para?.endTime) {
-      const endTimeComparison = a.para.endTime.localeCompare(b.para.endTime)
-      if (endTimeComparison !== 0) return endTimeComparison
-    } else if (a.para?.endTime) {
-      return -1
-    } else if (b.para?.endTime) {
-      return 1
-    }
-
-    const priorityA = a.para?.priority ?? 0
-    const priorityB = b.para?.priority ?? 0
-    if (priorityA !== priorityB) {
-      return priorityB - priorityA
-    }
-
-    const titleA = a.para?.title?.toLowerCase() ?? ''
-    const titleB = b.para?.title?.toLowerCase() ?? ''
-    return titleA.localeCompare(titleB)
-  })
-
-  const limit = 20
-  const itemsToShow = filteredItems.slice(0, limit)
-
-  const filteredOut = items.length ? items.length - itemsToShow.length : items.length - itemsToShow.length
-  const limitApplied = (items.length ?? 0) > itemsToShow.length
-
-  if (filteredOut > 0) {
-    itemsToShow.push({
-      ID: `${section.ID}-Filter`,
-      itemType: 'filterIndicator',
-      para: {
-        content: `There are also ${filteredOut} ${priorityFilteringHappening ? 'lower-priority' : ''} items currently hidden`,
-        filename: '',
-        type: 'text',
-        noteType: 'Notes',
-        rawContent: '',
-        priority: -1,
-      },
-    })
-  }
+  const hideSection = !items.length || (sharedSettings && sharedSettings[`${section.showSettingName}`] === false)
+  const sectionIsRefreshing = Array.isArray(pluginData.refreshing) && pluginData.refreshing.includes(section.sectionCode)
 
   let descriptionToUse = section.description
   if (descriptionToUse.includes('{count}')) {
@@ -170,12 +107,6 @@ const Section = ({ section }: SectionProps): React$Node => {
     descriptionToUse = descriptionToUse.replace('{totalCount}', `<span id='section${section.ID}TotalCount'}>${String(filteredOut)}</span>`)
   }
 
-  //----------------------------------------------------------------------
-  // Render
-  //----------------------------------------------------------------------
-  const hideSection = !items.length || (sharedSettings && sharedSettings[`${section.showSettingName}`] === false)
-  const sectionIsRefreshing = Array.isArray(pluginData.refreshing) && pluginData.refreshing.includes(section.sectionCode)
-
   return hideSection ? null : (
     <div className="section">
       <div className="sectionInfo">
@@ -186,12 +117,13 @@ const Section = ({ section }: SectionProps): React$Node => {
         </div>{' '}
         <div className="sectionDescription" dangerouslySetInnerHTML={{ __html: descriptionToUse }}></div>
         <div className="sectionButtons">
-          {buttons}
-          {section.sectionItems.length && section.sectionCode !== "PROJ" && enableInteractiveProcessing && (
-            <><button className="PCButton" onClick={handleInteractiveProcessingClick} title="Interactively process tasks one at a time">
-              <i className="fa-solid fa-arrows-rotate" style={{ opacity: 0.7 }}></i>
-              <span className="fa-layers-text" data-fa-transform="shrink-8" style={{ fontWeight: 500, paddingLeft: "3px" }}>{itemsToShow.length}</span>
-            </button>
+          {section.actionButtons?.map((item, index) => <CommandButton key={index} button={item} />) ?? []}
+          {section.sectionItems.length && section.sectionCode !== 'PROJ' && sharedSettings.enableInteractiveProcessing && (
+            <>
+              <button className="PCButton" onClick={handleInteractiveProcessingClick} title="Interactively process tasks one at a time">
+                <i className="fa-solid fa-arrows-rotate" style={{ opacity: 0.7 }}></i>
+                <span className="fa-layers-text" data-fa-transform="shrink-8" style={{ fontWeight: 500, paddingLeft: '3px' }}>{itemsToShow.length}</span>
+              </button>
             </>
           )}
         </div>