+++ jgclark.DashboardReact/src/react/customHooks/useInteractiveProcessing.jsx
@@ -1,7 +1,8 @@
 // useInteractiveProcessing.jsx
 // @flow
 /**
- * Custom hook for handling overdue processing logic.
+ * Custom hook for handling interactive processing logic
+ * Is loaded by all sections but should only run for the button that was checked
  * 
  * @param {Array<TSectionItem>} items - The items in the current section.
  * @param {TSection} thisSection - The current section information.
@@ -28,53 +29,55 @@ function useInteractiveProcessing(
   sharedSettings: TSharedSettings,
 ): void {
 
-  // Initialization effect. Sets the items to be interactively processed when reactSettings.interactiveProcessing has been set
+  // Check if this section should process interactively
+  const shouldProcess = reactSettings?.interactiveProcessing?.sectionName === thisSection.name
+
+  // Initialization effect
   useEffect(() => {
+    if (!shouldProcess) return
     if (!reactSettings) return
     if (!reactSettings.interactiveProcessing) return
     if (reactSettings.interactiveProcessing.sectionName !== thisSection.name) return
-    // logDebug('useInteractiveProcessing', `reactSettings.interactiveProcessing: ${reactSettings?.interactiveProcessing ? 'yes' : 'no'}; items in section ${String(thisSection.name)}: ${items.length} items, itemsCopy (remaining items): ${itemsCopy.length} items`)
+    // logDebug('useInteractiveProcessing', `reactSettings.interactiveProcessing: ${reactSettings?.interactiveProcessing ? 'yes' : 'no'}; items in section ${String(thisSection.name)}: ${items.length} items, itemsCopy (remaining items): ${itemsCopy.length} items`);
     if (reactSettings?.interactiveProcessing && items.length > 0 && reactSettings.interactiveProcessing.startingUp) {
-      logDebug('useInteractiveProcessing', `startingUp==true,Initializing itemsCopy to: ${items.length} items`)
+      logDebug('useInteractiveProcessing', `startingUp==true, Initializing itemsCopy to: ${items.length} items`)
       setItemsCopy([...items])
       setReactSettings((prev) => ({ ...prev, interactiveProcessing: { ...prev.interactiveProcessing, startingUp: false } }))
-    } else {
-      // logDebug('useInteractiveProcessing', `No initialization necessary...continuing`)
     }
-  }, [reactSettings?.interactiveProcessing, reactSettings, items, itemsCopy, thisSection.sectionCode, setItemsCopy])
+  }, [reactSettings, items, itemsCopy, thisSection, setItemsCopy, setReactSettings, shouldProcess])
 
   // Remove last processed item from itemsCopy
   useEffect(() => {
+    if (!shouldProcess) return
     if (!reactSettings) return
     if (!reactSettings.interactiveProcessing) return
+    if (reactSettings.interactiveProcessing.sectionName !== thisSection.name) return
     const { interactiveProcessing, dialogData } = reactSettings
     const { startingUp, currentIPIndex, totalTasks } = interactiveProcessing
     const dialogIsOpen = dialogData?.isOpen || false
-    logDebug('useInteractiveProcessing', `currentIPIndex: ${currentIPIndex}, dialogIsOpen: ${dialogIsOpen} itemsCopy.length: ${itemsCopy.length}`, itemsCopy[0])
     if (!dialogIsOpen && currentIPIndex < totalTasks && !startingUp) {
-      logDebug('useInteractiveProcessing', `Opening next overdue item at index: ${currentIPIndex + 1}`)
+      logDebug('useInteractiveProcessing', `dialog was closed; Slicing first item off itemsCopy; currentIPIndex:${currentIPIndex}/${totalTasks} dialogIsOpen:${dialogIsOpen} itemsCopy.length: ${itemsCopy.length}`, itemsCopy[0])
       // remove the first item off itemsCopy
       setItemsCopy([...itemsCopy.slice(1)])
     }
-  }, [itemsCopy, reactSettings, setItemsCopy])
+  }, [itemsCopy, reactSettings, setItemsCopy, shouldProcess])
 
   // Processing effect
   // Sets the next item in itemsCopy to be processed and opens the dialog
   useEffect(() => {
+    if (!shouldProcess) return
     if (!reactSettings) return
     if (!reactSettings.interactiveProcessing) return
     const { interactiveProcessing, dialogData } = reactSettings
     const { startingUp, sectionName, currentIPIndex, totalTasks } = interactiveProcessing
-    // logDebug('useInteractiveProcessing', `thisSection.name=${thisSection.name} / interactiveProcessing sectionName=${sectionName}`)
     if (sectionName !== thisSection.name) return
     if (startingUp) return
-    logDebug('useInteractiveProcessing', `${startingUp ? 'startingUp==true STOPPING' : 'startingUp==false'} ${JSP(interactiveProcessing)} itemsCopy.length=${itemsCopy.length}`)
+    logDebug('useInteractiveProcessing', `continuing... Section:${interactiveProcessing.sectionName}" ${interactiveProcessing.currentIPIndex}/${interactiveProcessing.totalTasks}} itemsCopy.length=${itemsCopy.length}`)
     if (itemsCopy.length === 0) return
-    logDebug('useInteractiveProcessing', `overdue processing, check for items to process, dialogData?.isOpen=${dialogData?.isOpen}`, reactSettings, interactiveProcessing)
     const dialogIsOpen = dialogData?.isOpen || false
 
     if (!dialogIsOpen && currentIPIndex < totalTasks && itemsCopy[0]) {
-      logDebug('useInteractiveProcessing', `Opening next overdue item at index: ${0}`)
+      logDebug('useInteractiveProcessing', `Opening next item: ${JSP(itemsCopy[0])}`)
       setReactSettings((prev) => ({
         ...prev,
         dialogData: {
@@ -101,7 +104,7 @@ function useInteractiveProcessing(
         setItemsCopy([])
       }
     }
-  }, [thisSection.sectionCode, reactSettings, itemsCopy, setItemsCopy, setReactSettings, sendActionToPlugin, sharedSettings])
+  }, [thisSection, reactSettings, itemsCopy, setItemsCopy, setReactSettings, sendActionToPlugin, sharedSettings, shouldProcess])
 }
 
 export default useInteractiveProcessing